Put defined homotopies in separate files
